refactor: split routes/orgs.rs (2026 lines) into domain submodules
Break the monolithic orgs.rs into focused files under routes/orgs/:
- mod.rs: shared types, helpers, auth, validation, macro, test helpers
- crud.rs: create_org, list_orgs, get_org
- teams.rs: create_team, list_teams, add_member, list_members, assign_wallet, list_team_wallets
- api_keys.rs: create_api_key, list_api_keys, revoke_api_key
- audit.rs: list_audit_logs
- budget.rs: set/get team/wallet budgets
- analytics.rs: get_team_stats, get_org_stats

All handlers remain pub and re-exported from mod.rs so external
references (lib.rs) continue to work unchanged. Tests distributed
to their respective submodules with shared helpers in mod.rs.
